requirements.txt
torch
opencv-contrib-python
numpy
scipy
Pillow
tensorflow
git+https://github.com/sedthh/pyxelate.git
scikit-image